fix(parse_tree): return error when missing section
Change `parseTree.get` to return an error if a section with the given
name does not exist. Prior to this, empty section slices were getting
created. This proved problematic as the decoder made assumptions about
the number of sections in a slice. Returning an explicit error in the
case of an absent section name makes the decoder more aware of when it
is safe to subscript a section slice. The decoder `decode` function also
now only attempts to get a section from the parse tree when it is going
to decode into a struct or slice of structs. The `TestDecode` case
includes a pointer to a struct, permitting a struct to be present in the
destination type, but missing from the decoded data.